add python script to estimate the basic reproduction number R of COVID 19 for different countries
this uses the data provided by [John Hopkins University](https://github.com/CSSEGISandData/COVID-19)

Ironically, countries which are featured in higher detail in the data
set (US, UK), do not work properly with the script because the data
set does not seem to provide aggregate data for these countries. If
this script turns out to be useful and somebody is interested in these
countries, I'll add some special code to deal with them.