add scripts to list all countries featured in the data and extract curves for all of them
there are issues for some of them, the biggest is probably South Korea
because this is the only country which is listed with a comma in its
name ("Korea, South") which breaks the simplistic data extraction
scheme used by the estimation script. We work around this by replacing
these country names by something which we can handle.