change the "Diamond Princess Estimate" to using new instead of cases
the intention is to give an rough indication of how well a country's
testing and reporting regime is currently working assuming that the
number of fatalities are correctly reported. (Using the total rate
would yield an indicator for how well the country did overall so far.)

also, the smoother is improved to not extrapolate data so easily
anymore. (this lead to spurious oscillations for the last few days of
the smoothened DPE where there is no actual data available yet.)