web ui: some improvements to the smoother code, use backward smoothing for the factor estimates
the smoother can be now switched between centralized and backward
smoothing and the detection of the smoothing range is now implemented
properly. For the R-factor estimates and the DPE-to-cases factors it
is more important that the specified number of sample points are
considered in the smoothened result to avoid wild oscillations for the
latest few days. Also the issue of the "backward smoothened" curves
trailing the raw curves is less pronounced there than for the case or
death curves...