web ui: use a crystal ball for the last few days of the prediction
the problem is that the cases reported in the immediate future have an
effect on the present. since the immediate future for the last few
reported dates is not yet known, we take the average number of new
cases of the last week where data is available and extrapolate with
that. if the number of new cases reported in the next few days are
close to the number used for extrapolation, the results for the last
few days should not change as much.

The effect of this does not seem to be very dramatic, though.