web ui: calculate the smoothened R on the smoothened new cases
to compensate, the result is not smoothened anymore. the reason for
this is that using the unprocessed numbers leads to computational
artifacts:

consider a country which only reports new cases every other day, i.e.,
which reports zero new cases every other day. Now consider an
infectivity distribution which also skips every second day. if we're
unlucky, we will calculate zero for every day. Although this effect
seems to affect the curves somewhat, the differences are rather small
in practice. anyway...