use adaptive smoothing
initially, central smoothing is used, which transitions to backward
smoothing for the last 14 days. the rationale is that central
smoothing requires a few data points from the future in order to
produce reasonable results, but for the last few days, these are
obviously not available yet, so the results for them show artifacts.