extract all countries in one go
before this, the database was read from scratch for each country,
while now it is read only a single time. This caused a timeout of the
cron script which updates the database every night. Since the whole
thing is now about two orders of magnitude faster, this should not
happen in the forseeable future anymore...